+++ packages/riverpod/CHANGELOG.md
@@ -1,6 +1,8 @@
 ## Unreleased major
 
-- **Breaking**: `AsyncValue` is now "sealed" (cannot be extended/implemented)
+- **Breaking**: `AsyncValue` is now "sealed" and `AsyncData/AsyncLoading/AsyncError`
+  are "final". This means that it is no-longer possible to subclass
+  `AsyncValue` or the associated classes.
 - Bumped minimum Dart SDK to >= 3.0.0-dev
 
 ## 2.3.6 - 2023-04-24